Update NeuBox shadows and PlayerPage icons to use theme colors
This commit modifies the NeuBox component to use `Theme.of(context).shadowColor` and `Theme.of(context).highlightColor` for its shadow effects.

Additionally, the skip previous and skip next icons in the PlayerPage now use `Theme.of(context).colorScheme.inverseSurface` for their active color and `Colors.grey.shade600` for their inactive color.

The marquee package has been added as a dependency and is used in FavouritesPage and HomePage to display scrolling song titles.